IMPROVE Code organization
